Publish window and workspace ownership on the control tree
