+++ agtermCore/Sources/agtermctlKit/SocketClient.swift
@@ -289,9 +289,14 @@ struct SocketClient {
         }.joined(separator: "\n")
     }
 
-    /// Render a tree as an indented workspace → session listing (no trailing newline).
+    /// Render a tree as an indented workspace → session listing (no trailing newline), under a header naming
+    /// the projected window. Every session node repeats that window id and its workspace id, which the
+    /// indentation already shows — only `--json` reads them per row.
     private static func formatTree(_ tree: ControlTree) -> String {
         var lines: [String] = []
+        if let windowId = tree.windowId {
+            lines.append("window [\(windowId)]")
+        }
         for workspace in tree.workspaces {
             let mark = workspace.active ? "*" : " "
             lines.append("\(mark) \(workspace.name)  [\(workspace.id)]")
